Initial commit with base project
This commit includes a working website with
testing. Functionality includes the ability to
search/filter by cocktail name, category or
ingredients in real time. The search displays as tiles on
the screen and the user is able to open each
tile into a dialog window with full details
on how to make the selected drink.